Checkpoint for tokio client feature gate
