+++ gneiss-mqtt/Cargo.toml
@@ -16,15 +16,12 @@ features = []
 all-features = true
 
 [features]
-default = ["tokio"] # remove soon
+default = [] # remove soon
 testing = [] # enables environmentally-controlled integration tests and long-running tests (30+ seconds) that check reconnect, backoff, etc...
 rustls = ["dep:rustls", "dep:rustls-native-certs", "dep:rustls-pemfile", "dep:rustls-pki-types", "dep:tokio-rustls"]
 native-tls = ["dep:native-tls", "dep:tokio-native-tls"]
 websockets = ["dep:http", "dep:tungstenite", "dep:tokio-tungstenite", "dep:stream-ws"]
 tokio = ["dep:tokio"]
-
-# Eventually: tokio, rustls, nativetls, async_std, threaded, http_proxy, websockets
-# Consider: mqtt311, mqtt5, log
 
 [dev-dependencies]
 assert_matches = "1.5.0"
